// src/models/tags.d.ts
export interface Datum {
    id:         number;
    attributes: Attributes;
}

export interface Attributes {
    Title:       string;
    createdAt:   Date;
    updatedAt:   Date;
    publishedAt: Date;
    slug:        string;
}